メゾット課題　再提出

+++ src/main/java/Method/q07/SumNumbers2.java
@@ -5,15 +5,15 @@ public abstract class SumNumbers2 {
 	public static void main(String[] args) {
 		int num1 = 5;
 		double num2 = 3.3;
-		double sum = calculateSum(num1, num2);
+		int sum = calculateSum(num1, num2);
 		System.out.println("第一引数(整数):" + num1);
 		System.out.println("第二引数(実数):" + num2);
 		System.out.println("加算結果:" + sum);
 
 	}
 
-	public static double calculateSum(int num1, double num2) {
-		double sum = num1 + num2;
+	public static int calculateSum(int num1, double num2) {
+		int sum = (int) (num1 + num2);
 		return sum;
 	}
